requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
autopep8==2.3.2
certifi==2025.4.26
charset-normalizer==3.4.2
dateparser==1.2.1
frozenlist==1.6.2
idna==3.10
multidict==6.4.4
numpy==2.2.6
pandas==2.3.0
propcache==0.3.1
pycodestyle==2.13.0
pycryptodome==3.23.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
six==1.17.0
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
